% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{acf2}
\alias{acf2}
\title{Calculate lag10 autocorrelation}
\usage{
acf2(x, lag = 10, ...)
}
\arguments{
\item{x}{numeric vector}

\item{lag}{integer}

\item{...}{additional arguments to \code{acf}}
}
\description{
A wrapper for the function acf that returns the autocorrelation for
the specified lag.  Missing values are removed.
}
\seealso{
\code{\link{acf}}
}

