% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hmm-methods.R
\name{baumWelchUpdate}
\alias{baumWelchUpdate}
\title{Function for updating parameters for emission probabilities}
\usage{
baumWelchUpdate(param, assay_list)
}
\arguments{
\item{param}{A container for the HMM parameters}

\item{assay_list}{list of log R ratios and B allele frequencies}
}
\description{
This function is not meant to be called directly by the user.  It
is exported in the package NAMESPACE for internal use by other BioC
packages.
}

