% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/defunct.R
\name{BeadStudioSetList}
\alias{BeadStudioSet}
\alias{BeadStudioSetList}
\alias{BeadStudioSetList,BeadStudioSet-method}
\alias{BeadStudioSetList,character-method}
\title{Deprecated BeadStudioSet and BeadStudioSetList}
\usage{
BeadStudioSetList(x, ...)

BeadStudioSet(filenames, header_info, genome = c("hg19", "hg18"),
  annotationPkg, chromosome = 1:22, ...)
}
\arguments{
\item{x}{}

\item{...}{}

\item{filenames}{full path to input files}

\item{header_info}{header information}

\item{genome}{character string indicating genome build}

\item{annotationPkg}{name of annotation package}

\item{chromosome}{which chromosomes to process}
}
\description{
The BeadStudioSet, BeadStudioSetList classes are deprecated.  The
corresponding constructor and classes are provided only for
compatability with older versions and will be defunct at the next
release.
}
\details{
Use the replacement function SnpArrayExperiment.  For reading in
plain text source files such as GenomeStudio-processed data
(Illumina platform), see \code{\link{CopyNumScanParams}} and
\code{\link{parseSourceFile}}.
}
\seealso{
\code{link{CopyNumScanParams}} \code{\link{parseSourceFile}}
}
\keyword{internal}

