% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{constrainMu2}
\alias{constrainMu2}
\alias{constrainSd2}
\alias{copyNumberLimits}
\title{Constraints for updating the means of the copy number states}
\usage{
constrainMu2(mu)

constrainSd2(sigma)

copyNumberLimits(is.log)
}
\arguments{
\item{mu}{numeric vector of means}

\item{sigma}{numeric vector of standard deviations (non-negative)}

\item{is.log}{whether copy number estimates are on log scale}
}
\value{
numeric vector of means

constrained standard deviations

numeric vector giving the lower and upper values of the restricted range
}
\description{
Constraints for updating the means of the copy number states

Constraints for updating the standard deviations of the BAFs

Restricted range for CN values
}
\examples{
copyNumberLimits(is.log=TRUE)
copyNumberLimits(is.log=FALSE)
}
\keyword{manip}

