% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{hmm}
\alias{hmm}
\alias{hmm,BafLrrSetList-method}
\alias{hmm,BeadStudioSet-method}
\alias{hmm,BeadStudioSetList-method}
\alias{hmm,SnpSet2-method}
\alias{hmm,oligoSetList-method}
\alias{hmm,oligoSnpSet-method}
\alias{robustSds}
\title{Deprecated functions in the VanillaICE package}
\usage{
hmm(object, ...)

robustSds(x, takeLog = FALSE, ...)
}
\arguments{
\item{object}{see \code{showMethods(hmm)} for a listing of classes for which this method is defined}

\item{...}{additional arguments to \code{mad}}

\item{x}{a numeric matrix}

\item{takeLog}{whether to first log2 transform the numeric matrix}
}
\value{
a matrix of the same dimension as the input. Within a column, the entries are identical
}
\description{
These functions have been deprecated.  The functions are only
provided only for compatability with older versions and will be
defunct at the next release.
}
\seealso{
\code{\link{mad}}
}

