% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hmm-methods.R
\name{doUpdate}
\alias{doUpdate}
\title{Helper function to determine whether to update the HMM parameters via the Baum-Welch algorithm}
\usage{
doUpdate(param)
}
\arguments{
\item{param}{An object containing parameters for the HMM}
}
\description{
This function is not intended to be called directly by the user,
and is exported only for internal use by other BioC packages.
}
\seealso{
\code{\link{HmmParam}}
}

