% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-ArrayViews.R
\name{dropSexChrom}
\alias{dropSexChrom}
\title{Filter sex chromosomes}
\usage{
dropSexChrom(object)
}
\arguments{
\item{object}{an object for which the methods \code{seqnames} and \code{rowRanges} are defined.}
}
\value{
an object of the same class as the input
}
\description{
Removes markers on chromosomes X and Y.
}

