% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{emission}
\alias{emission}
\alias{emission,HmmParam-method}
\alias{emission<-}
\alias{emission<-,HMM-method}
\alias{emission<-,HmmParam-method}
\title{Methods to set and get emission probabilities}
\usage{
emission(object)

emission(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(emission)}}

\item{value}{a matrix of emission probabilities}
}
\value{
matrix
}
\description{
Get or set a matrix of emission probabilities. This function is
exported primarily for internal use by other BioC packages.
}

