% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{emissionParam}
\alias{emissionParam}
\alias{emissionParam,HMM-method}
\alias{emissionParam,HmmGRanges-method}
\alias{emissionParam,HmmParam-method}
\alias{emissionParam<-}
\alias{emissionParam<-,HmmGRanges,EmissionParam-method}
\alias{emissionParam<-,HmmParam,EmissionParam-method}
\title{Accessor for parameters used to compute emission probabilities}
\usage{
emissionParam(object)

emissionParam(object) <- value
}
\arguments{
\item{object}{an object of class \code{EmissionParam}}

\item{value}{an object of class \code{EmissionParam}}
}
\value{
\code{\link{EmissionParam}} instance
}
\description{
Parameters for computing emission probabilities include the
starting values for the Baum Welch update and initial state
probabilities.
}
\examples{
hparam <- HmmParam()
emissionParam(hparam)
ep <- EmissionParam()
cn_means(ep) <- log2(c(.1/2, 1/2, 2/2, 2/2, 3/2, 4/2))
emissionParam(hparam) <- ep
}

