% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-HmmGRanges.R
\docType{methods}
\name{state,HmmGRanges-method}
\alias{state,HmmGRanges-method}
\title{Accessor for copy number state}
\usage{
\S4method{state}{HmmGRanges}(object)
}
\arguments{
\item{object}{a \code{HmmGRanges} object}
}
\description{
Extract the copy number state for each genomic interval.
}

