% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-HMMList.R
\name{HMMList}
\alias{HMMList}
\title{Constructor for \code{HMMList} class}
\usage{
HMMList(object)
}
\arguments{
\item{object}{a list. Each element of the list is in instance of
the \code{HMM} class.}
}
\description{
The constructor function for the \code{HMMList} class. The
constructor is useful for representing a list of \code{HMM}
objects.
}
\seealso{
\code{\linkS4class{HMMList}} \code{\linkS4class{HMM}} \code{\link{hmm2}}
}

