% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-HmmTrellisParam.R
\name{HmmTrellisParam}
\alias{HmmTrellisParam}
\title{Constructor for HmmTrellisParam class}
\usage{
HmmTrellisParam(ylimits = list(c(0, 1), c(-3, 1)), expandfun = function(g) {
      width(g) * 50 })
}
\arguments{
\item{ylimits}{length-two list of the y-axis limits for B allele
frequencies and log R ratios, respectively}

\item{expandfun}{a function that takes a length-one \code{GRanges}
object as an argument and computes a width relative to the width of
the \code{GRanges} object}
}
\description{
Constructor for HmmTrellisParam class
}

