% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\name{icePlatforms}
\alias{icePlatforms}
\title{List platforms for which ICE option is supported.}
\usage{
icePlatforms()
}
\value{
A character vector of the annotation packages that are supported for
 the ICE option
}
\description{
When procecessing genotypes with the \pkg{crlmm}, confidence
scores for the diallelic genotype calls are available.  One can
estimate the emission probabilities for the crlmm diallelic
genotypes using the confidence scores by setting the value of
\code{ICE} to TRUE in the constructor for the \code{HmmOptionList}
class.  Currently, only certain platforms are supported for this
option.
}
\examples{
icePlatforms()
}
\references{
Scharpf, RB et al., 2008, Annals of Applied Statistics
}
\keyword{misc}

