% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-IdiogramParams.R
\docType{methods}
\name{IdiogramParams}
\alias{IdiogramParams}
\alias{plot,IdiogramParams,ANY-method}
\alias{plot,IdiogramParams-method}
\title{Constructor for IdiogramParam objects}
\usage{
IdiogramParams(seqnames = character(), seqlengths = numeric(),
  unit = "kb", genome = "hg19", box = list(color = "blue", lwd = 1))

\S4method{plot}{IdiogramParams,ANY}(x, y, ...)
}
\arguments{
\item{seqnames}{length-one character vector providing chromosome name}

\item{seqlengths}{length-one numeric vector indicating size of chromosome}

\item{unit}{character string indicating unit for genomic position}

\item{genome}{character string indicating genome build}

\item{box}{a list of parameters for plotting the box around the part of the idiogram that is plotted}

\item{x}{an IdiogramParam object}

\item{y}{ignored}

\item{...}{ignored}
}
\value{
IdiogramParam object
}
\description{
Parameters for plotting idiograms
}

