% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ArrayViews.R
\docType{methods}
\name{lrrFile}
\alias{bafFile}
\alias{bafFile,ArrayViews-method}
\alias{gtFile}
\alias{gtFile,ArrayViews-method}
\alias{lrrFile}
\alias{lrrFile,ArrayViews-method}
\alias{lrrFile<-}
\alias{lrrFile<-,ArrayViews-method}
\title{Accessors for objects of class ArrayViews}
\usage{
lrrFile(object)

lrrFile(object) <- value

bafFile(object)

gtFile(object)

\S4method{lrrFile}{ArrayViews}(object)

\S4method{lrrFile}{ArrayViews}(object) <- value

\S4method{bafFile}{ArrayViews}(object)

\S4method{gtFile}{ArrayViews}(object)
}
\arguments{
\item{object}{see showMethods("lrrFile")}

\item{value}{a character vector of filenames for the log R ratios}
}
\description{
Accessors for objects of class ArrayViews
}
\examples{
views <- ArrayViews(parsedPath=tempdir())
sourcePaths(views)
lrrFile(views)
bafFile(views)
gtFile(views)
}

