% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ArrayViews.R, R/methods-SnpArrayExperiment.R
\docType{methods}
\name{isHeterozygous}
\alias{isHeterozygous}
\alias{isHeterozygous,ArrayViews-method}
\alias{isHeterozygous,SnpArrayExperiment-method}
\alias{isHeterozygous,matrix-method}
\alias{isHeterozygous,numeric-method}
\title{Assess whether genotype is heterozygous based on BAFs}
\usage{
isHeterozygous(object, cutoff)

\S4method{isHeterozygous}{ArrayViews}(object, cutoff)

\S4method{isHeterozygous}{SnpArrayExperiment}(object, cutoff)

\S4method{isHeterozygous}{numeric}(object, cutoff)

\S4method{isHeterozygous}{matrix}(object, cutoff)
}
\arguments{
\item{object}{a SnpArrayExperiment or ArrayViews object containing
BAFs, a matrix of BAFs, or a numeric vector of BAFs.
vector of BAFs}

\item{cutoff}{a length-two numeric vector providing the range of
BAFs consistent with allelic  heterozygosity}
}
\description{
Assess whether genotype is heterozygous based on BAFs
}
\examples{
snp_exp <- getExampleSnpExperiment()
is_het <- isHeterozygous(snp_exp[, 1], c(0.4, 0.6))
table(is_het)
}

