% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-LogLik.R
\name{LogLik}
\alias{LogLik}
\title{Constructor for LogLik class}
\usage{
LogLik(loglik = numeric(), tolerance = 1L)
}
\arguments{
\item{loglik}{length-one numeric vector for the log likelihood of the Viterbi state path}

\item{tolerance}{if the difference in the log-likelihood of the Viterbi state path after the Baum-Welch update is less than the specified tolerance, no additional Baum-Welch updates are required}
}
\description{
A container for the log likelihood of the Viterbi state path.
Stores the log likelihood from succesive updates of model
parameters.  When the difference between the log likelihoods at
iteration i and i-1 is below the tolerance, no additional updates
are performed.
}
\seealso{
\linkS4class{LogLik}
}

