% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ArrayViews.R, R/methods-SnpArrayExperiment.R
\docType{methods}
\name{genotypes}
\alias{baf,ArrayViews-method}
\alias{baf,SnpArrayExperiment-method}
\alias{copyNumber,SnpArrayExperiment-method}
\alias{genotypes}
\alias{genotypes,ArrayViews-method}
\alias{genotypes,SnpArrayExperiment-method}
\alias{lrr,ArrayViews-method}
\alias{lrr,SnpArrayExperiment-method}
\title{Accessor for SNP genotypes}
\usage{
genotypes(object)

\S4method{lrr}{ArrayViews}(object)

\S4method{baf}{ArrayViews}(object)

\S4method{genotypes}{ArrayViews}(object)

\S4method{baf}{SnpArrayExperiment}(object)

\S4method{copyNumber}{SnpArrayExperiment}(object)

\S4method{lrr}{SnpArrayExperiment}(object)

\S4method{genotypes}{SnpArrayExperiment}(object)
}
\arguments{
\item{object}{see \code{showMethods("genotypes")}}
}
\description{
Extract SNP genotypes. Genotypes are assumed to be represented as
integers: 1=AA, 2=AB, 3=BB.
}
\seealso{
\code{copyNumber}
}

