% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{numberFeatures}
\alias{numberFeatures}
\alias{numberFeatures,FilterParam-method}
\alias{numberFeatures,HMM-method}
\alias{numberFeatures,HmmGRanges-method}
\title{The number of SNP/nonpolymorphic probes contained in a genomic interval}
\usage{
numberFeatures(object)
}
\arguments{
\item{object}{see \code{showMethods(numberFeatures)}}
}
\description{
The number of SNP/nonpolymorphic probes contained in a genomic interval
}

