% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\docType{methods}
\name{[[,oligoSetList,ANY,ANY-method}
\alias{[[,oligoSetList,ANY,ANY-method}
\title{Subset method for deprecated \code{oligoSetList}}
\usage{
\S4method{[[}{oligoSetList,ANY,ANY}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{x}{a \code{oligoSetList} object}

\item{i}{A length-one numeric vector specifying which chromosome to extract}

\item{j}{A numeric vector specifying samples to extract (optional)}

\item{...}{Ignored}

\item{exact}{Ignored}
}
\description{
The \code{oligoSetList} class is deprecated.  Use
\code{\link{SnpArrayExperiment}} instead.
}

