% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-HmmGRanges.R, R/methods-HmmTrellisParam.R
\docType{methods}
\name{xyplotList}
\alias{xygrid}
\alias{xyplotList}
\alias{xyplotList,GRangesList,SnpArrayExperiment-method}
\alias{xyplotList,HmmGRanges,SnpArrayExperiment-method}
\title{Lattice-style plots for granges and SnpArrayExperiment objects}
\usage{
xyplotList(granges, se, param = HmmTrellisParam())

\S4method{xyplotList}{HmmGRanges,SnpArrayExperiment}(granges, se,
  param = HmmTrellisParam())

\S4method{xyplotList}{GRangesList,SnpArrayExperiment}(granges, se,
  param = HmmTrellisParam())

xygrid(trellis_plot, viewports, granges)
}
\arguments{
\item{granges}{a \code{HmmGRanges} object}

\item{se}{a \code{SnpArrayExperiment}}

\item{param}{trellis parameters for plotting HMM}

\item{trellis_plot}{an object of class \code{trellis}}

\item{viewports}{a list of viewports as provided by the \code{viewports} function}
}
\description{
Data for the graphic is generated by a call to \code{grangesData}.
}
\examples{
snp_exp <- getExampleSnpExperiment()
seqlevels(snp_exp, force=TRUE) <- "chr22"
fit <- hmm2(snp_exp)
g <- reduce(hemizygous(fit), min.gapwidth=500e3)
trellis_param <- HmmTrellisParam()
fig <- xyplotList(g, snp_exp, trellis_param)
vps <- viewports()
xygrid(fig[[1]], vps, g)
}
\seealso{
\code{\link{viewports}}
}

