% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{probability}
\alias{probability}
\title{Accessor for probability filter}
\usage{
probability(object)
}
\arguments{
\item{object}{a \code{FilterParam} object}
}
\description{
Accessor for probability filter
}

