% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{segs}
\alias{segs}
\alias{segs,HMM-method}
\title{Accessor for the HMM segments}
\usage{
segs(object)
}
\arguments{
\item{object}{see \code{showMethods(segs)}}
}
\value{
a \code{GRanges}-derived object
}
\description{
Accessor to obtain all segments from the HMM.
}

