% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{snp_exp}
\alias{snp_exp}
\title{An example SnpArrayExperiment}
\format{a \code{SnpArrayExperiment} object}
\description{
A container for low-level summaries used for downstream copy number
estimation, including log R ratios, B allele frequencies, and
genotypes
}
\keyword{datasets}

