% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ArrayViews.R
\docType{methods}
\name{SnpExperiment}
\alias{SnpExperiment}
\alias{SnpExperiment,ArrayViews-method}
\title{Constructor for SnpArrayExperiment}
\usage{
SnpExperiment(object)

\S4method{SnpExperiment}{ArrayViews}(object)
}
\arguments{
\item{object}{see \code{showMethods('SnpExperiment')} for a list of supported objects}
}
\description{
A single-argument generic function to construct a SnpArrayExperiment.
}
\examples{
view <- ArrayViews()
SnpExperiment(view)
}

