% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sourcePaths}
\alias{sourcePaths}
\alias{sourcePaths,ArrayViews-method}
\title{Accessor for file paths containing SNP-level summaries}
\usage{
sourcePaths(object)
}
\arguments{
\item{object}{an \code{ArrayViews} object}
}
\description{
Files containing SNP-level summaries for log R ratios, B allele
frequencies, and genotypes -- one sample per subject -- are
required.
}
\examples{
sourcePaths(ArrayViews())
}

