% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmm-methods.R
\docType{methods}
\name{start,oligoSnpSet-method}
\alias{start,oligoSnpSet-method}
\title{Retrieve genomic location of SNPs}
\usage{
\S4method{start}{oligoSnpSet}(x)
}
\arguments{
\item{x}{a \code{oligoSnpSet} object}
}
\description{
Retrieve genomic location of SNPs
}

