% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-Viterbi.R
\docType{methods}
\name{state-methods}
\alias{state,Viterbi-method}
\alias{state-methods}
\title{Accessor for the Viterbi state path}
\usage{
\S4method{state}{Viterbi}(object)
}
\arguments{
\item{object}{a \code{Viterbi} object}
}
\description{
The states are represented as integers: 1=homozygous deletion,
2=hemizygous deletion, 3=diploid normal heterozygosity, 4=diploid
region of homozygosity, 5=single copy gain, 6=two or more copy gain.
}

