% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{threshold}
\alias{threshold}
\title{Threshold numeric values}
\usage{
threshold(x, lim = c(-Inf, Inf), amount = 0)
}
\arguments{
\item{x}{numeric matrix or vector}

\item{lim}{limit at which to threshold entries in \code{x}}

\item{amount}{see \code{\link{jitter}}}
}
\description{
Threshold numeric values according to user-specific limits.  The
thresholded values can also be jittered near the limits.
}
\examples{
x <- rnorm(1000, 0, 3)
y <- threshold(x, c(-5,5))
range(y)
}
\seealso{
\code{\link{jitter}}
}

