% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hmm-methods.R
\name{updateHmmParams}
\alias{updateHmmParams}
\title{Run the Baum-Welch algorithm to update HMM parameters}
\usage{
updateHmmParams(object, emission_param = EmissionParam(),
  transition_param = TransitionParam())
}
\arguments{
\item{object}{a \code{\link{SnpArrayExperiment}} object}

\item{emission_param}{a \code{\link{EmissionParam}} object}

\item{transition_param}{a \code{\link{TransitionParam}} object}
}
\description{
This function is not intended to be called directly by the user. It
is exported in the package NAMESPACE for internal use by other BioC
packages.
}

