% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{VanillaICE}
\alias{VanillaICE}
\alias{VanillaICE-package}
\title{A hidden markov model for detection of germline copy number variants from arrays}
\description{
A hidden markov model for detection of germline copy number variants from arrays
}

