% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-HmmTrellisParam.R
\name{viewports}
\alias{viewports}
\title{Default viewports for plotting CNV data with lattice-style graphics}
\usage{
viewports()
}
\value{
list
}
\description{
Default viewports for plotting CNV data with lattice-style graphics
}
\examples{
vps <- viewports()
}
\seealso{
\code{\link{xyplotList}} \code{\link{xygrid}}
}

