% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-Viterbi.R
\docType{methods}
\name{show,Viterbi-method}
\alias{show,Viterbi-method}
\title{Show method for objects of class \code{Viterbi}}
\usage{
\S4method{show}{Viterbi}(object)
}
\arguments{
\item{object}{a \code{Viterbi} object}
}
\description{
Show method for objects of class \code{Viterbi}
}

