% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\name{viterbi2Wrapper}
\alias{viterbi2Wrapper}
\title{Wrapper function for fitting the viterbi algorithm}
\usage{
viterbi2Wrapper(index.samples, cnStates, prOutlierBAF = list(initial = 1e-05,
  max = 0.001, maxROH = 1e-05), p.hom = 0.05, is.log, limits,
  normalIndex = 3L, nupdates = 10, tolerance = 5, computeLLR = TRUE,
  returnEmission = FALSE, verbose = FALSE, grFun, matrixFun, snp.index,
  anyNP)
}
\arguments{
\item{index.samples}{Index for the samples that are to be processed.}

\item{cnStates}{\code{numeric} vector for the initial copy number state means.}

\item{prOutlierBAF}{A list with elements 'initial', 'max', and 'maxROH' corresponding to the initial estimate of the probability that a B allele frequency (BAF) is an outlier, the maximum value for this parameter over states that do not involve homozygous genotypes, and the maximum value over states that assume homozygous genotypes.  This parameter is experimental and could be used to fine tune the HMM for different platforms. For example, the BAFs for the Affy platform are typically more noisey than the BAFs for Illumina.  One may want to set small values of these parameters for Illumina (e.g, 1e-5, 1e-3, and 1e-5) and larger values for Affy (e.g., 1e-3, 0.01, 1e-3).}

\item{p.hom}{\code{numeric}: weight for observing homozygous genotypes.  For value \code{0}, homozygous genotypes / B allele frequencies have the same emission probability in the 'normal' state as in the states hemizygous deletion and in copy-neutral region of homozygosity.  Regions of homozygosity are common in normal genomes. For small values of \code{p.hom}, hemizygous deletions will only be called if the copy number estimates show evidence of a decrease from normal.}

\item{is.log}{\code{logical}: Whether the copy number estimates in the \code{r} matrix are on the log-scale.}

\item{limits}{\code{numeric} vector of length two specifying the range of the copy number estimates in \code{r}. Values of \code{r} outside of this range are truncated. See \code{copyNumberLimits}.}

\item{normalIndex}{\code{integer} specifying the index for the normal state.  Note that states must be ordered by the mean of the copy number state. E.g., state 1 is homozygous deletion (0 copies), state 2 is hemizygous deletion (1 copy), normal (2 copies), ...  In a 6-state HMM, normalIndex should be 3.}

\item{nupdates}{\code{integer} specifying the maximum number of iterations for reestimating the mean and variance for each of the copy number states.  The number of iterations may be fewer than \code{nupdates} if the difference in the log-likelihood between successive iterations is less than \code{tolerance}.}

\item{tolerance}{\code{numeric} value for indicating convergence of the log-likelihood. If the difference in the log-likelihood of the observed data given the HMM model at iteration i and i-1 is less than tolerance, no additional updates of model parameters using the EM algorithm is needed.}

\item{computeLLR}{Logical.  Whether to compute a log likelihood ratio (LLR) comparing the predicted state to the normal state. This is calculated post-hoc and is not precisely the likelihood estimated from the Viterbi algorithm. When \code{FALSE}, the LLR is not calculated and the algorithm is slightly faster.}

\item{returnEmission}{Logical. If TRUE, an array of emission probabilities are returned.  The dimensions of the array are SNPs, samples, and copy number states.}

\item{verbose}{Logical. Whether to print some of the details of the processing.}

\item{grFun}{An R function for coercing the state-path from the HMM
to a GRanges object. Takes advantage of lexical scope.}

\item{matrixFun}{An R function for subsetting the assay data (takes
advantage of lexical scope).}

\item{snp.index}{The SNP indices}

\item{anyNP}{An indicator for whether any of the markers are
nonpolymorphic, and therefore BAFs / genotypes are ignored}
}
\value{
A \code{GRanges} object if \code{returnEmission} is FALSE.  Otherwise,
 an array of emission probabilities is returned.
}
\description{
The viterbi algorithm, implemented in C, estimates the optimal
state path as well as the forward and backward variables that are
used for updating the mean and variances in a copy number HMM.  The
function \code{viterbi2Wrapper} should not be called directly by
the user. Rather, users should fit the HMM by passing an
appropriate container to the method \code{hmm}.  We document the
\code{viterbi2Wrapper} arguments as several of the arguments can be
modified from their default value when passed from the \code{hmm}
method through the \code{...}. In particular, \code{nupdates},
\code{p.hom}, and \code{prOutlierBaf}.
}
\keyword{smooth}

